move card-form from cards to forms as index. enable delete buttons on contracts and profile
