+++ models/user.js
@@ -15,16 +15,16 @@ module.exports = (sequelize, DataTypes) => {
 
   User.init({
     auth0_id: { type: DataTypes.STRING, allowNull: false, unique: true },
-    name: DataTypes.STRING,
-    number: DataTypes.STRING,
-    email: DataTypes.STRING,
-    street1: DataTypes.STRING,
-    street2: DataTypes.STRING,
-    city: DataTypes.STRING,
-    state: DataTypes.STRING,
-    post_zip: DataTypes.STRING,
-    country: DataTypes.STRING,
-    picture: DataTypes.STRING,
+    name: { type: DataTypes.STRING, defaultValue: '', allowNull: false, unique: false },
+    number: { type: DataTypes.STRING, defaultValue: '', allowNull: false, unique: false },
+    email: { type: DataTypes.STRING, defaultValue: '', allowNull: false, unique: true },
+    street1: { type: DataTypes.STRING, defaultValue: '', allowNull: false, unique: false },
+    street2: { type: DataTypes.STRING, defaultValue: '', allowNull: true, unique: false },
+    city: { type: DataTypes.STRING, defaultValue: '', allowNull: false, unique: false },
+    state: { type: DataTypes.STRING, defaultValue: '', allowNull: false, unique: false },
+    post_zip: { type: DataTypes.STRING, defaultValue: '', allowNull: false, unique: false },
+    country: { type: DataTypes.STRING, defaultValue: '', allowNull: false, unique: false },
+    picture: { type: DataTypes.STRING, defaultValue: 'https://res.cloudinary.com/ryanbuckleyca/image/upload/c_scale,h_150,w_150/v1600109993/user_bgu0at.jpg', allowNull: false, unique: false },
     address: { type: DataTypes.VIRTUAL, get() {
         // function also exists in User model
         const fields = ['name', 'street1', 'street2', 'city', 'state', 'post_zip', 'country']
