+++ src/views/contract.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { withAuth0 } from '@auth0/auth0-react';
-import CardForm from '../components/cards/card-form';
+import CardForm from '../components/forms';
 import CardFormTopSideContract from '../components/forms/contract-topside';
 import CardFormFieldsContract from '../components/forms/contract-main';
